feat: add bridge tests
