+++ tasks/deployment/accounts.ts
@@ -3,7 +3,7 @@ import { task } from 'hardhat/config'
 import { cliOpts } from '../../cli/defaults'
 import { updateItemValue, writeConfig } from '../../cli/config'
 
-task('migrate:accounts', '[localhost] Creates protocol accounts and saves them in graph config')
+task('migrate:accounts', 'Creates protocol accounts and saves them in graph config')
   .addParam('graphConfig', cliOpts.graphConfig.description, cliOpts.graphConfig.default)
   .setAction(async (taskArgs, hre) => {
     const { graphConfig, getDeployer } = hre.graph({
